Add base entity to switchbee
